+++ pkg/internal/provider/docs/doc_helpers_test.go
@@ -1,4 +1,4 @@
-package helpers
+package docs
 
 import (
 	"testing"
