+++ pkg/internal/provider/special_values.go
@@ -1,0 +1,33 @@
+package provider
+
+import (
+	"fmt"
+)
+
+const (
+	BooleanTrue    = "true"
+	BooleanFalse   = "false"
+	BooleanDefault = "default"
+
+	IntDefault       = -1
+	IntDefaultString = "-1"
+)
+
+func booleanStringFromBool(value bool) string {
+	if value {
+		return BooleanTrue
+	} else {
+		return BooleanFalse
+	}
+}
+
+func BooleanStringToBool(value string) (bool, error) {
+	switch value {
+	case BooleanTrue:
+		return true, nil
+	case BooleanFalse:
+		return false, nil
+	default:
+		return false, fmt.Errorf("cannot retrieve boolean value from %s", value)
+	}
+}
