+++ pkg/internal/provider/validators/validators_test.go
@@ -1,4 +1,4 @@
-package helpers
+package validators
 
 import (
 	"testing"
